Code review feedback

+++ src/devices/machine/dp83932c.h
@@ -208,8 +208,9 @@ private:
 	u16 m_reg[64];
 	u64 m_cam[16];
 
-	u32 read_bus_word(offs_t address);
-	void write_bus_word(offs_t address, u32 data);
+	// These wrappers are needed because some SONIC data is stored as 16-bit data stored on dword boundaries when in 32-bit mode
+	u16 read_bus_word(offs_t address);
+	void write_bus_word(offs_t address, u16 data);
 };
 
 DECLARE_DEVICE_TYPE(DP83932C, dp83932c_device)
